requirements.txt
django
social-auth-app-django
whitenoise
pymysql
requests
opensearch-py
python-dotenv
django-modeltranslation
django-opensearch-dsl
Pillow
djangorestframework
markdown
django-filter
django-rest-knox
rest-social-auth
